Cleanup tests and comments
- Update minimum Go version to go 1.19
